requirements.txt
pandas==1.0.1
psycopg2==2.8.6
SQLAlchemy_Utils==0.38.2
numpy==1.19.5
tqdm==4.62.3
geopy==2.2.0
Flask==2.0.1
SQLAlchemy==1.4.31
python-dotenv==0.19.2
